Move test code around
